@@ -1,6 +1,6 @@
 module github.com/grafana/phlare
 
-go 1.18
+go 1.19
 
 require (
 	github.com/bufbuild/connect-go v1.4.1
