@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"hash/fnv"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"strings"
 	"sync"
@@ -75,7 +74,7 @@ func (tg *TargetGroup) sync(groups []*targetgroup.Group) {
 	var actives []*Target
 	tg.droppedTargets = []*Target{}
 	for _, group := range groups {
-		targets, dropped, err := tg.targetsFromGroup(group)
+		targets, dropped, err := tg.TargetsFromGroup(group)
 		if err != nil {
 			level.Error(tg.logger).Log("msg", "creating targets failed", "err", err)
 			continue
@@ -85,9 +84,7 @@ func (tg *TargetGroup) sync(groups []*targetgroup.Group) {
 				actives = append(actives, t)
 			}
 		}
-		for _, dt := range dropped {
-			tg.droppedTargets = append(tg.droppedTargets, dt)
-		}
+		tg.droppedTargets = append(tg.droppedTargets, dropped...)
 	}
 
 	for _, t := range actives {
@@ -246,7 +243,7 @@ func (t *Target) fetchProfile(ctx context.Context, profileType string, buf io.Wr
 	}
 	defer resp.Body.Close()
 
-	b, err := ioutil.ReadAll(io.TeeReader(resp.Body, buf))
+	b, err := io.ReadAll(io.TeeReader(resp.Body, buf))
 	if err != nil {
 		return fmt.Errorf("failed to read body: %w", err)
 	}
