@@ -5,91 +5,234 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"time"
 
-	grpchealth "github.com/bufbuild/connect-grpchealth-go"
 	"github.com/felixge/fgprof"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/kv/codec"
 	"github.com/grafana/dskit/kv/memberlist"
 	"github.com/grafana/dskit/ring"
+	"github.com/grafana/dskit/runtimeconfig"
 	"github.com/grafana/dskit/services"
 	grpcgw "github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/pkg/errors"
-	"google.golang.org/genproto/googleapis/api/httpbody"
-	"google.golang.org/protobuf/encoding/protojson"
-	"gopkg.in/yaml.v2"
-
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/collectors"
 	"github.com/prometheus/common/version"
 	"github.com/thanos-io/thanos/pkg/discovery/dns"
 	"github.com/weaveworks/common/middleware"
 	"github.com/weaveworks/common/server"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
+	"google.golang.org/genproto/googleapis/api/httpbody"
+	"google.golang.org/protobuf/encoding/protojson"
+	"gopkg.in/yaml.v2"
 
 	agentv1 "github.com/grafana/phlare/api/gen/proto/go/agent/v1"
 	"github.com/grafana/phlare/api/gen/proto/go/agent/v1/agentv1connect"
-	ingesterv1connect "github.com/grafana/phlare/api/gen/proto/go/ingester/v1/ingesterv1connect"
+	"github.com/grafana/phlare/api/gen/proto/go/ingester/v1/ingesterv1connect"
 	"github.com/grafana/phlare/api/gen/proto/go/push/v1/pushv1connect"
 	"github.com/grafana/phlare/api/gen/proto/go/querier/v1/querierv1connect"
 	statusv1 "github.com/grafana/phlare/api/gen/proto/go/status/v1"
 	"github.com/grafana/phlare/api/openapiv2"
 	"github.com/grafana/phlare/pkg/agent"
+	"github.com/grafana/phlare/pkg/api"
 	"github.com/grafana/phlare/pkg/distributor"
+	"github.com/grafana/phlare/pkg/frontend"
+	"github.com/grafana/phlare/pkg/frontend/frontendpb/frontendpbconnect"
 	"github.com/grafana/phlare/pkg/ingester"
 	objstoreclient "github.com/grafana/phlare/pkg/objstore/client"
 	"github.com/grafana/phlare/pkg/objstore/providers/filesystem"
 	phlarecontext "github.com/grafana/phlare/pkg/phlare/context"
 	"github.com/grafana/phlare/pkg/querier"
+	"github.com/grafana/phlare/pkg/querier/worker"
+	"github.com/grafana/phlare/pkg/scheduler"
+	"github.com/grafana/phlare/pkg/scheduler/schedulerpb/schedulerpbconnect"
 	"github.com/grafana/phlare/pkg/usagestats"
 	"github.com/grafana/phlare/pkg/util"
 	"github.com/grafana/phlare/pkg/util/build"
+	"github.com/grafana/phlare/pkg/validation"
+	"github.com/grafana/phlare/pkg/validation/exporter"
 )
 
 // The various modules that make up Phlare.
 const (
-	All          string = "all"
-	Agent        string = "agent"
-	Distributor  string = "distributor"
-	Server       string = "server"
-	Ring         string = "ring"
-	Ingester     string = "ingester"
-	MemberlistKV string = "memberlist-kv"
-	Querier      string = "querier"
-	GRPCGateway  string = "grpc-gateway"
-	Storage      string = "storage"
-	UsageReport  string = "usage-stats"
+	All               string = "all"
+	Agent             string = "agent"
+	Distributor       string = "distributor"
+	Server            string = "server"
+	Ring              string = "ring"
+	Ingester          string = "ingester"
+	MemberlistKV      string = "memberlist-kv"
+	Querier           string = "querier"
+	GRPCGateway       string = "grpc-gateway"
+	Storage           string = "storage"
+	UsageReport       string = "usage-stats"
+	QueryFrontend     string = "query-frontend"
+	QueryScheduler    string = "query-scheduler"
+	RuntimeConfig     string = "runtime-config"
+	Overrides         string = "overrides"
+	OverridesExporter string = "overrides-exporter"
 
-	// RuntimeConfig            string = "runtime-config"
-	// Overrides                string = "overrides"
-	// OverridesExporter        string = "overrides-exporter"
-	// TenantConfigs            string = "tenant-configs"
-	// IngesterQuerier          string = "ingester-querier"
-	// QueryFrontend            string = "query-frontend"
 	// QueryFrontendTripperware string = "query-frontend-tripperware"
-	// RulerStorage             string = "ruler-storage"
-	// Ruler                    string = "ruler"
-	// TableManager             string = "table-manager"
 	// Compactor                string = "compactor"
 	// IndexGateway             string = "index-gateway"
 	// IndexGatewayRing         string = "index-gateway-ring"
-	// QueryScheduler           string = "query-scheduler"
 )
 
 var objectStoreTypeStats = usagestats.NewString("store_object_type")
 
-func (f *Phlare) initQuerier() (services.Service, error) {
-	q, err := querier.New(f.Cfg.Querier, f.ring, nil, f.logger, f.auth)
+func (f *Phlare) initQueryFrontend() (services.Service, error) {
+	if f.Cfg.Frontend.Addr == "" {
+		addr, err := util.GetFirstAddressOf(f.Cfg.Frontend.InfNames)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to get frontend address")
+		}
+
+		f.Cfg.Frontend.Addr = addr
+	}
+
+	if f.Cfg.Frontend.Port == 0 {
+		f.Cfg.Frontend.Port = f.Cfg.Server.HTTPListenPort
+	}
+
+	frontendSvc, err := frontend.NewFrontend(f.Cfg.Frontend, log.With(f.logger, "component", "frontend"), f.reg)
 	if err != nil {
 		return nil, err
 	}
-	// Those API are not meant to stay but allows us for testing through Grafana.
-	f.Server.HTTP.Handle("/pyroscope/render", http.HandlerFunc(q.RenderHandler))
-	f.Server.HTTP.Handle("/pyroscope/label-values", http.HandlerFunc(q.LabelValuesHandler))
-	querierv1connect.RegisterQuerierServiceHandler(f.Server.HTTP, q, f.auth)
+	querierv1connect.RegisterQuerierServiceHandler(f.Server.HTTP, querier.NewGRPCRoundTripper(frontendSvc), f.auth)
+	frontendpbconnect.RegisterFrontendForQuerierHandler(f.Server.HTTP, frontendSvc, f.auth)
+	return frontendSvc, nil
+}
 
-	return q, nil
+func (f *Phlare) initRuntimeConfig() (services.Service, error) {
+	if len(f.Cfg.RuntimeConfig.LoadPath) == 0 {
+		// no need to initialize module if load path is empty
+		return nil, nil
+	}
+
+	f.Cfg.RuntimeConfig.Loader = loadRuntimeConfig
+
+	// make sure to set default limits before we start loading configuration into memory
+	validation.SetDefaultLimitsForYAMLUnmarshalling(f.Cfg.LimitsConfig)
+
+	serv, err := runtimeconfig.New(f.Cfg.RuntimeConfig, prometheus.WrapRegistererWithPrefix("phlare_", f.reg), log.With(f.logger, "component", "runtime-config"))
+	if err == nil {
+		// TenantLimits just delegates to RuntimeConfig and doesn't have any state or need to do
+		// anything in the start/stopping phase. Thus we can create it as part of runtime config
+		// setup without any service instance of its own.
+		f.TenantLimits = newTenantLimits(serv)
+	}
+
+	f.RuntimeConfig = serv
+
+	f.Server.HTTP.Methods("GET").Path("/runtime_config").Handler(runtimeConfigHandler(f.RuntimeConfig, f.Cfg.LimitsConfig))
+	f.Server.HTTP.Methods("GET").Path("/api/v1/tenant_limits").Handler(middleware.AuthenticateUser.Wrap(validation.TenantLimitsHandler(f.Cfg.LimitsConfig, f.TenantLimits)))
+	f.IndexPage.AddLinks(api.RuntimeConfigWeight, "Current runtime config", []api.IndexPageLink{
+		{Desc: "Entire runtime config (including overrides)", Path: "/runtime_config"},
+		{Desc: "Only values that differ from the defaults", Path: "/runtime_config?mode=diff"},
+	})
+	return serv, err
+}
+
+func (f *Phlare) initOverrides() (serv services.Service, err error) {
+	f.Overrides, err = validation.NewOverrides(f.Cfg.LimitsConfig, f.TenantLimits)
+	// overrides don't have operational state, nor do they need to do anything more in starting/stopping phase,
+	// so there is no need to return any service.
+	return nil, err
+}
+
+func (f *Phlare) initOverridesExporter() (services.Service, error) {
+	overridesExporter, err := exporter.NewOverridesExporter(
+		f.Cfg.OverridesExporter,
+		&f.Cfg.LimitsConfig,
+		f.TenantLimits,
+		log.With(f.logger, "component", "overrides-exporter"),
+		f.reg,
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to instantiate overrides-exporter")
+	}
+	if f.reg != nil {
+		f.reg.MustRegister(overridesExporter)
+	}
+
+	f.Server.HTTP.Methods("GET", "POST").Path("/overrides-exporter/ring").HandlerFunc(overridesExporter.RingHandler)
+	f.IndexPage.AddLinks(api.DefaultWeight, "Overrides-exporter", []api.IndexPageLink{
+		{Desc: "Ring status", Path: "/overrides-exporter/ring"},
+	})
+	return overridesExporter, nil
+}
+
+func (f *Phlare) initQueryScheduler() (services.Service, error) {
+	f.Cfg.QueryScheduler.ServiceDiscovery.SchedulerRing.ListenPort = f.Cfg.Server.HTTPListenPort
+
+	s, err := scheduler.NewScheduler(f.Cfg.QueryScheduler, f.Overrides, log.With(f.logger, "component", "scheduler"), f.reg)
+	if err != nil {
+		return nil, errors.Wrap(err, "query-scheduler init")
+	}
+	schedulerpbconnect.RegisterSchedulerForFrontendHandler(f.Server.HTTP, s)
+	schedulerpbconnect.RegisterSchedulerForQuerierHandler(f.Server.HTTP, s)
+	return s, nil
+}
+
+// setupWorkerTimeout sets the max loop duration for the querier worker and frontend worker
+// to 90% of the read or write http timeout, whichever is smaller.
+// This is to ensure that the worker doesn't timeout before the http handler and that the connection
+// is refreshed.
+func (f *Phlare) setupWorkerTimeout() {
+	timeout := f.Cfg.Server.HTTPServerReadTimeout
+	if f.Cfg.Server.HTTPServerWriteTimeout < timeout {
+		timeout = f.Cfg.Server.HTTPServerWriteTimeout
+	}
+
+	if timeout > 0 {
+		f.Cfg.Worker.MaxLoopDuration = time.Duration(float64(timeout) * 0.9)
+		f.Cfg.Frontend.MaxLoopDuration = time.Duration(float64(timeout) * 0.9)
+	}
+}
+
+func (f *Phlare) initQuerier() (services.Service, error) {
+	querierSvc, err := querier.New(f.Cfg.Querier, f.ring, nil, log.With(f.logger, "component", "querier"), f.auth)
+	if err != nil {
+		return nil, err
+	}
+	if !f.isModuleActive(QueryFrontend) {
+		querierv1connect.RegisterQuerierServiceHandler(f.Server.HTTP, querierSvc, f.auth)
+	}
+	worker, err := worker.NewQuerierWorker(f.Cfg.Worker, querier.NewGRPCHandler(querierSvc), log.With(f.logger, "component", "querier-worker"), f.reg)
+	if err != nil {
+		return nil, err
+	}
+
+	f.Server.HTTP.Handle("/pyroscope/render", util.AuthenticateUser(f.Cfg.MultitenancyEnabled).Wrap(http.HandlerFunc(querierSvc.RenderHandler)))
+	f.Server.HTTP.Handle("/pyroscope/label-values", util.AuthenticateUser(f.Cfg.MultitenancyEnabled).Wrap(http.HandlerFunc(querierSvc.LabelValuesHandler)))
+
+	sm, err := services.NewManager(querierSvc, worker)
+	if err != nil {
+		return nil, err
+	}
+	w := services.NewFailureWatcher()
+	w.WatchManager(sm)
+
+	return services.NewBasicService(func(ctx context.Context) error {
+		err := sm.StartAsync(ctx)
+		if err != nil {
+			return err
+		}
+		return sm.AwaitHealthy(ctx)
+	}, func(ctx context.Context) error {
+		select {
+		case <-ctx.Done():
+			return nil
+		case err := <-w.Chan():
+			return err
+		}
+	}, func(failureCase error) error {
+		sm.StopAsync()
+		return sm.AwaitStopped(context.Background())
+	}), nil
 }
 
 func (f *Phlare) getPusherClient() pushv1connect.PusherServiceClient {
@@ -112,7 +255,8 @@ func (f *Phlare) initGRPCGateway() (services.Service, error) {
 }
 
 func (f *Phlare) initDistributor() (services.Service, error) {
-	d, err := distributor.New(f.Cfg.Distributor, f.ring, nil, f.reg, f.logger, f.auth)
+	f.Cfg.Distributor.DistributorRing.ListenPort = f.Cfg.Server.HTTPListenPort
+	d, err := distributor.New(f.Cfg.Distributor, f.ring, nil, f.Overrides, f.reg, log.With(f.logger, "component", "distributor"), f.auth)
 	if err != nil {
 		return nil, err
 	}
@@ -121,11 +265,16 @@ func (f *Phlare) initDistributor() (services.Service, error) {
 	f.pusherClient = d
 
 	pushv1connect.RegisterPusherServiceHandler(f.Server.HTTP, d, f.auth)
+	f.Server.HTTP.Path("/distributor/ring").Methods("GET", "POST").Handler(d)
+	f.IndexPage.AddLinks(api.DefaultWeight, "Distributor", []api.IndexPageLink{
+		{Desc: "Ring status", Path: "/distributor/ring"},
+	})
+
 	return d, nil
 }
 
 func (f *Phlare) initAgent() (services.Service, error) {
-	a, err := agent.New(&f.Cfg.AgentConfig, f.logger, f.getPusherClient)
+	a, err := agent.New(&f.Cfg.AgentConfig, log.With(f.logger, "component", "agent"), f.getPusherClient)
 	if err != nil {
 		return nil, err
 	}
@@ -158,17 +307,30 @@ func (f *Phlare) initMemberlistKV() (services.Service, error) {
 
 	f.MemberlistKV = memberlist.NewKVInitService(&f.Cfg.MemberlistKV, f.logger, dnsProvider, f.reg)
 
+	f.Cfg.Distributor.DistributorRing.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
 	f.Cfg.Ingester.LifecyclerConfig.RingConfig.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
+	f.Cfg.QueryScheduler.ServiceDiscovery.SchedulerRing.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
+	f.Cfg.OverridesExporter.Ring.KVStore.MemberlistKV = f.MemberlistKV.GetMemberlistKV
 
+	f.Cfg.Frontend.QuerySchedulerDiscovery = f.Cfg.QueryScheduler.ServiceDiscovery
+	f.Cfg.Worker.QuerySchedulerDiscovery = f.Cfg.QueryScheduler.ServiceDiscovery
+
+	f.Server.HTTP.Path("/memberlist").Handler(api.MemberlistStatusHandler("", f.MemberlistKV))
+	f.IndexPage.AddLinks(api.MemberlistWeight, "Memberlist", []api.IndexPageLink{
+		{Desc: "Status", Path: "/memberlist"},
+	})
 	return f.MemberlistKV, nil
 }
 
 func (f *Phlare) initRing() (_ services.Service, err error) {
-	f.ring, err = ring.New(f.Cfg.Ingester.LifecyclerConfig.RingConfig, "ingester", "ring", f.logger, prometheus.WrapRegistererWithPrefix("phlare_", f.reg))
+	f.ring, err = ring.New(f.Cfg.Ingester.LifecyclerConfig.RingConfig, "ingester", "ring", log.With(f.logger, "component", "ring"), prometheus.WrapRegistererWithPrefix("phlare_", f.reg))
 	if err != nil {
 		return nil, err
 	}
 	f.Server.HTTP.Path("/ring").Methods("GET", "POST").Handler(f.ring)
+	f.IndexPage.AddLinks(api.DefaultWeight, "Ingester", []api.IndexPageLink{
+		{Desc: "Ring status", Path: "/ring"},
+	})
 	return f.ring, nil
 }
 
@@ -202,24 +364,33 @@ func (f *Phlare) context() context.Context {
 func (f *Phlare) initIngester() (_ services.Service, err error) {
 	f.Cfg.Ingester.LifecyclerConfig.ListenPort = f.Cfg.Server.HTTPListenPort
 
-	ingester, err := ingester.New(f.context(), f.Cfg.Ingester, f.Cfg.PhlareDB, f.storageBucket)
+	ingester, err := ingester.New(f.context(), f.Cfg.Ingester, f.Cfg.PhlareDB, f.storageBucket, f.Overrides)
 	if err != nil {
 		return nil, err
 	}
-	prefix, handler := grpchealth.NewHandler(grpchealth.NewStaticChecker(ingesterv1connect.IngesterServiceName))
-	f.Server.HTTP.NewRoute().PathPrefix(prefix).Handler(handler)
 	ingesterv1connect.RegisterIngesterServiceHandler(f.Server.HTTP, ingester, f.auth)
 	return ingester, nil
 }
 
 func (f *Phlare) initServer() (services.Service, error) {
-	prometheus.MustRegister(version.NewCollector("phlare"))
+	f.reg.MustRegister(version.NewCollector("phlare"))
+	f.reg.Unregister(collectors.NewGoCollector())
+	// register collector with additional metrics
+	f.reg.MustRegister(collectors.NewGoCollector(
+		collectors.WithGoCollectorRuntimeMetrics(collectors.MetricsAll),
+	))
 	DisableSignalHandling(&f.Cfg.Server)
 	f.Cfg.Server.Registerer = prometheus.WrapRegistererWithPrefix("phlare_", f.reg)
 	// Not all default middleware works with http2 so we'll add then manually.
 	// see https://github.com/grafana/phlare/issues/231
 	f.Cfg.Server.DoNotAddDefaultHTTPMiddleware = true
 
+	f.setupWorkerTimeout()
+	if f.isModuleActive(QueryScheduler) {
+		// to ensure that the query scheduler is always able to handle the request, we need to double the timeout
+		f.Cfg.Server.HTTPServerReadTimeout = 2 * f.Cfg.Server.HTTPServerReadTimeout
+		f.Cfg.Server.HTTPServerWriteTimeout = 2 * f.Cfg.Server.HTTPServerWriteTimeout
+	}
 	serv, err := server.New(f.Cfg.Server)
 	if err != nil {
 		return nil, err
@@ -276,6 +447,26 @@ func (f *Phlare) initServer() (services.Service, error) {
 		return nil, err
 	}
 
+	// register static assets
+	f.Server.HTTP.PathPrefix("/static/").Handler(http.FileServer(http.FS(api.StaticFiles)))
+
+	// register index page
+	f.IndexPage = api.NewIndexPageContent()
+	f.Server.HTTP.Path("/").Handler(api.IndexHandler("", f.IndexPage))
+
+	// Add index page links.
+	f.IndexPage.AddLinks(api.BuildInfoWeight, "Build information", []api.IndexPageLink{
+		{Desc: "Build information", Path: "/api/v1/status/buildinfo"},
+	})
+	f.IndexPage.AddLinks(api.ConfigWeight, "Current config", []api.IndexPageLink{
+		{Desc: "Including the default values", Path: "/api/v1/status/config"},
+		{Desc: "Only values that differ from the defaults", Path: "/api/v1/status/config/diff"},
+		{Desc: "Default values", Path: "/api/v1/status/config/default"},
+	})
+	f.IndexPage.AddLinks(api.OpenAPIDefinitionWeight, "OpenAPI definition", []api.IndexPageLink{
+		{Desc: "Swagger JSON", Path: "/api/swagger.json"},
+	})
+
 	return s, nil
 }
 
@@ -319,7 +510,6 @@ func (f *Phlare) initUsageReport() (services.Service, error) {
 
 type statusService struct {
 	statusv1.UnimplementedStatusServiceServer
-	configYaml    string
 	defaultConfig *Config
 	actualConfig  *Config
 }
