@@ -4,10 +4,10 @@ import (
 	"context"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 
 	phlarecontext "github.com/grafana/phlare/pkg/phlare/context"
-	query "github.com/grafana/phlare/pkg/phlaredb/query"
+	"github.com/grafana/phlare/pkg/phlaredb/query"
+	"github.com/grafana/phlare/pkg/util"
 )
 
 type contextKey uint8
@@ -18,12 +18,10 @@ const (
 )
 
 type headMetrics struct {
-	head *Head
-
-	series        prometheus.GaugeFunc
+	series        prometheus.Gauge
 	seriesCreated *prometheus.CounterVec
 
-	profiles        prometheus.GaugeFunc
+	profiles        prometheus.Gauge
 	profilesCreated *prometheus.CounterVec
 
 	sizeBytes   *prometheus.GaugeVec
@@ -35,68 +33,64 @@ type headMetrics struct {
 
 func newHeadMetrics(reg prometheus.Registerer) *headMetrics {
 	m := &headMetrics{
-		seriesCreated: promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+		seriesCreated: prometheus.NewCounterVec(prometheus.CounterOpts{
 			Name: "phlare_tsdb_head_series_created_total",
 			Help: "Total number of series created in the head",
 		}, []string{"profile_name"}),
-		rowsWritten: promauto.With(reg).NewCounterVec(
+		rowsWritten: prometheus.NewCounterVec(
 			prometheus.CounterOpts{
 				Name: "phlare_rows_written",
 				Help: "Number of rows written to a parquet table.",
 			},
 			[]string{"type"}),
-		profilesCreated: promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+		profilesCreated: prometheus.NewCounterVec(prometheus.CounterOpts{
 			Name: "phlare_head_profiles_created_total",
 			Help: "Total number of profiles created in the head",
 		}, []string{"profile_name"}),
-		sampleValuesIngested: promauto.With(reg).NewCounterVec(
+		sampleValuesIngested: prometheus.NewCounterVec(
 			prometheus.CounterOpts{
 				Name: "phlare_head_ingested_sample_values_total",
 				Help: "Number of sample values ingested into the head per profile type.",
 			},
 			[]string{"profile_name"}),
-		sampleValuesReceived: promauto.With(reg).NewCounterVec(
+		sampleValuesReceived: prometheus.NewCounterVec(
 			prometheus.CounterOpts{
 				Name: "phlare_head_received_sample_values_total",
 				Help: "Number of sample values received into the head per profile type.",
 			},
 			[]string{"profile_name"}),
-
-		// this metric is not registered using promauto, as it has a callback into the header
 		sizeBytes: prometheus.NewGaugeVec(
 			prometheus.GaugeOpts{
 				Name: "phlare_head_size_bytes",
 				Help: "Size of a particular in memory store within the head phlaredb block.",
 			},
 			[]string{"type"}),
+		series: prometheus.NewGauge(prometheus.GaugeOpts{
+			Name: "phlare_tsdb_head_series",
+			Help: "Total number of series in the head block.",
+		}),
+		profiles: prometheus.NewGauge(prometheus.GaugeOpts{
+			Name: "phlare_head_profiles",
+			Help: "Total number of profiles in the head block.",
+		}),
 	}
 
-	// metrics that call into the head
-	m.series = promauto.With(reg).NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "phlare_tsdb_head_series",
-		Help: "Total number of series in the head block.",
-	}, func() float64 {
-		if m.head == nil {
-			return 0.0
-		}
-		return float64(m.head.index.totalSeries.Load())
-	})
-	m.profiles = promauto.With(reg).NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "phlare_head_profiles",
-		Help: "Total number of profiles in the head block.",
-	}, func() float64 {
-		if m.head == nil {
-			return 0.0
-		}
-		return float64(m.head.index.totalProfiles.Load())
-	})
-
-	if reg != nil {
-		reg.MustRegister(
-			m,
-		)
-	}
+	m.register(reg)
 	return m
+}
+
+func (m *headMetrics) register(reg prometheus.Registerer) {
+	if reg == nil {
+		return
+	}
+	m.series = util.RegisterOrGet(reg, m.series)
+	m.seriesCreated = util.RegisterOrGet(reg, m.seriesCreated)
+	m.profiles = util.RegisterOrGet(reg, m.profiles)
+	m.profilesCreated = util.RegisterOrGet(reg, m.profilesCreated)
+	m.sizeBytes = util.RegisterOrGet(reg, m.sizeBytes)
+	m.rowsWritten = util.RegisterOrGet(reg, m.rowsWritten)
+	m.sampleValuesIngested = util.RegisterOrGet(reg, m.sampleValuesIngested)
+	m.sampleValuesReceived = util.RegisterOrGet(reg, m.sampleValuesReceived)
 }
 
 func contextWithHeadMetrics(ctx context.Context, m *headMetrics) context.Context {
@@ -111,24 +105,6 @@ func contextHeadMetrics(ctx context.Context) *headMetrics {
 	return m
 }
 
-func (m *headMetrics) setHead(head *Head) *headMetrics {
-	m.head = head
-	return m
-}
-
-func (m *headMetrics) Describe(ch chan<- *prometheus.Desc) {
-	m.sizeBytes.Describe(ch)
-}
-
-func (m *headMetrics) Collect(ch chan<- prometheus.Metric) {
-	if m.head != nil {
-		for _, t := range m.head.tables {
-			m.sizeBytes.WithLabelValues(t.Name()).Set(float64(t.Size()))
-		}
-	}
-	m.sizeBytes.Collect(ch)
-}
-
 type blocksMetrics struct {
 	query *query.Metrics
 
@@ -136,13 +112,15 @@ type blocksMetrics struct {
 }
 
 func newBlocksMetrics(reg prometheus.Registerer) *blocksMetrics {
-	return &blocksMetrics{
+	m := &blocksMetrics{
 		query: query.NewMetrics(reg),
-		blockOpeningLatency: promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
+		blockOpeningLatency: prometheus.NewHistogram(prometheus.HistogramOpts{
 			Name: "phlaredb_block_opening_duration",
 			Help: "Latency of opening a block in seconds",
 		}),
 	}
+	m.blockOpeningLatency = util.RegisterOrGet(reg, m.blockOpeningLatency)
+	return m
 }
 
 func contextWithBlockMetrics(ctx context.Context, m *blocksMetrics) context.Context {
