@@ -19,9 +19,7 @@ import (
 )
 
 func TestQueryIndex(t *testing.T) {
-	head := newTestHead(t)
-
-	a, err := newProfileIndex(32, newHeadMetrics(prometheus.NewRegistry()).setHead(head.Head))
+	a, err := newProfileIndex(32, newHeadMetrics(prometheus.NewRegistry()))
 	require.NoError(t, err)
 
 	for j := 0; j < 10; j++ {
@@ -54,7 +52,7 @@ func TestQueryIndex(t *testing.T) {
 	}
 
 	tmpFile := t.TempDir() + "/test.db"
-	err = a.WriteTo(context.Background(), tmpFile)
+	_, err = a.writeTo(context.Background(), tmpFile)
 	require.NoError(t, err)
 
 	r, err := index.NewFileReader(tmpFile)
